Introductory Data Analysis Using Python
Collaborative project conducted in an introductory programming course attempting to predict the electrical power output of a micro gas turbine. The goal of this project is to forecast the electrical power output (el_power) of a 3-kilowatt commercial micro gas turbine using the input control signal (input_voltage) over time. This study investigates the relationship between the input signal and energy output through the use of time-series data and regression models. The outcome helps to optimize turbine operations and increase energy system efficiency.